CCUPSY_officer/requirements.txt
ArticutAPI>=1.0.0
beautifulsoup4
python-dotenv>=0.19.0
lxml>=4.9.0
pandas>=1.3.0
requests>=2.26.0
openpyxl>=3.0.0
discord
discord.py>=2.0.0
pathlib>=1.0.0
importlib>=1.0.0
argparse>=1.4.0
pprint>=0.1